+++ talker/example/talker_example.dart
@@ -1,7 +1,23 @@
 import 'package:talker/talker.dart';
 
+class HttpTalkerLog extends TalkerLog {
+  HttpTalkerLog(String message) : super(message);
+
+  @override
+  AnsiPen get pen => AnsiPen()..xterm(49);
+
+  @override
+  String generateTextMessage() {
+    return pen.write(message);
+  }
+}
+
 Future<void> main() async {
-  await Talker.instance.configure();
+  await Talker.instance.configure(
+    settings: const TalkerSettings(
+      registeredTypes: [HttpTalkerLog],
+    ),
+  );
 
   try {
     throw Exception('Test service exception');
@@ -16,22 +32,21 @@ Future<void> main() async {
   }
 
   Talker.instance.log(
-    'Big app crashing exception in Some service',
+    'Server error',
     logLevel: LogLevel.critical,
     additional: {
-      "timestamp": 1510417124782,
       "status": 500,
       "error": "Internal Server Error",
-      "exception": "com.netflix.hystrix.exception.HystrixRuntimeException",
-      "message":
-          "ApplicationRepository#save(Application) failed and no fallback available.",
-      "path": "/application"
     },
   );
 
-  Talker.instance.fine('Log info');
-  Talker.instance.error('Log info');
-  Talker.instance.good('Log info');
-  Talker.instance.verbose('Log info');
-  Talker.instance.warning('Log info');
+  Talker.instance.fine('Log fine');
+  Talker.instance.error('Log error');
+  Talker.instance.good('Log good');
+  Talker.instance.verbose('Log verbose');
+  Talker.instance.warning('Log warning');
+  Talker.instance.critical('Log critical');
+
+  final httpLog = HttpTalkerLog('Http good');
+  Talker.instance.logTyped(httpLog);
 }