+++ talker/lib/src/models/talker_data/talker_exception.dart
@@ -1,5 +1,7 @@
 import 'package:talker/talker.dart';
 
+/// Base implementation of [TalkerDataInterface]
+/// to handle ONLY [Exceptions]s
 class TalkerException implements TalkerDataInterface {
   TalkerException(
     this.exception, {
@@ -11,31 +13,41 @@ class TalkerException implements TalkerDataInterface {
     _time = time ?? DateTime.now();
   }
 
+  late DateTime _time;
+
+  /// {@macro talker_data_exception}
   @override
   final Exception exception;
 
+  /// {@macro talker_data_message}
   @override
   final String? message;
 
-  @override
-  final Error? error = null;
-
+  /// {@macro talker_data_stackTrace}
   @override
   final StackTrace? stackTrace;
 
-  @override
-  final Map<String, dynamic>? additional = null;
-
+  /// {@macro talker_data_loglevel}
   @override
   final LogLevel? logLevel;
 
+  /// {@macro talker_data_generateTextMessage}
   @override
   String generateTextMessage() {
-    return '$titleText$displayMessage$displayException$displayStackTrace';
+    return '$displayTitle$displayMessage$displayException$displayStackTrace';
   }
 
+  /// {@macro talker_data_time}
   @override
   DateTime get time => _time;
 
-  late DateTime _time;
+  /// {@macro talker_data_error}
+  /// Not used in [TalkerException]
+  @override
+  final Error? error = null;
+
+  /// {@macro talker_data_additional}
+  /// Not used in [TalkerException]
+  @override
+  final Map<String, dynamic>? additional = null;
 }