+++ talker_error_handler/lib/src/models/error_level/error_level.dart
@@ -1,3 +1,8 @@
+/// {@template errorLevel}
+/// [ErrorLevel] [errorLevel] - field to understand the degree of error
+/// How critical is it
+/// {@endtemplate}
+
 enum ErrorLevel {
   debug,
   warning,