+++ talker/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.5.0
+- Add all docs
+- Update core packages versions to 0.5.0
+- Fix package issues
+
+## 0.4.0
+- Add logTyped method
+- Add documentation
+- Updated settings and working
+
 ## 0.3.0
 - Update settings cnfiguration
 - Update internal logic