+++ talker_logger/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.5.0
+- Release 0.5.0 talker package
+- Update README.md
+
 ## 0.4.0
 - Update logs base formating and coloring